fix(PAN-653): move git-activity.ts to src/lib and fix cross-layer imports
Move git-activity service from src/dashboard/server/services/ to src/lib/
to eliminate lib→dashboard coupling. The old location now re-exports from
the canonical src/lib path for backward compatibility.

Also fix reopen.ts to import review-status from src/lib instead of the
dashboard re-export path.

Update all test mocks and dynamic imports to reference the new path.
